+++ Scripts/AttackScripts/HitData.cs
@@ -7,7 +7,7 @@ public struct HitData
 	public Vector2 Skb {get; set;}
 	public Vector2 Vkb {get; set;}
 	public float Damage {get; set;}
-	public int Stun {get; set;}
+	public float Stun {get; set;}
 	public int Hitpause {get; set;}
 	public Hitbox Hitter {get; set;}
 	public Hurtbox Hitee {get; set;}
@@ -35,7 +35,7 @@ public struct HitData
 		}
 	}
 	
-	public HitData(Vector2 skb, Vector2 vkb, float damage, int stun, int hitpause, Hitbox hitter, Hurtbox hitee)
+	public HitData(Vector2 skb, Vector2 vkb, float damage, float stun, int hitpause, Hitbox hitter, Hurtbox hitee)
 	{
 		Skb = skb;
 		Vkb = vkb;
