+++ Scripts/AttackScripts/IAttacker.cs
@@ -5,7 +5,7 @@ public interface IAttacker
 {
 	float DamageDoneMult{get; set;}
 	float KnockbackDoneMult{get; set;}
-	int StunDoneMult{get; set;}
+	float StunDoneMult{get; set;}
 	int TeamNumber{get; set;}
 	
 	bool CanHit(IHittable hitObject);
